+++ src/webpack/webpack.base.config.js
@@ -2,14 +2,15 @@ import path from 'path'
 import moduleConfig from './module'
 import resolveConfig from './resolve'
 import externalsConfig from './externals'
-import optimizationConfig from './optimization'
+// import optimizationConfig from './optimization'
 const cwdPath = process.cwd() // 工程项目root path
 
 export default function(parentDirPath) {
     //
     return {
         mode: 'production',
-        // devtool: 'source-map',
+        // devtool: 'eval', // webpack5属性
+        target: ['web', 'es5'], // webpack5属性
         entry: {},
         output: {
             path: path.resolve(cwdPath, './build'),
@@ -18,7 +19,7 @@ export default function(parentDirPath) {
             publicPath: './'
         },
         //打包多文件和公共模块配置
-        optimization: optimizationConfig,
+        // optimization: optimizationConfig, // webpack5下部分被废弃
         resolveLoader: {
             modules: [
                 path.join(cwdPath, 'node_modules'),
@@ -28,6 +29,11 @@ export default function(parentDirPath) {
         },
         externals: externalsConfig,
         module: moduleConfig(),
-        resolve: resolveConfig
+        resolve: resolveConfig,
+        experiments: {
+            asyncWebAssembly: true, // 开启 webpack5 的 asyncWebAssembly 功能
+            syncWebAssembly: true,
+            topLevelAwait: true // 开启 top-level-await
+        }
     }
 }