+++ src/webpack/webpack.test.config.js
@@ -1,23 +1,23 @@
-import path from 'path';
-import webpack from 'webpack';
+import path from 'path'
+import webpack from 'webpack'
 // import pluginsConfig from './plugins';
-import { merge } from 'webpack-merge';
-import webpackBaseConfig from './webpack.base.config';
-import SpeedMeasurePlugin from 'speed-measure-webpack-plugin';
-import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
-const cwdPath = process.cwd(); // 工程项目root path
+import { merge } from 'webpack-merge'
+import webpackBaseConfig from './webpack.base.config'
+import SpeedMeasurePlugin from 'speed-measure-webpack-plugin'
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
+const cwdPath = process.cwd() // 工程项目root path
 const analyzerConfig = {
     analyzerMode: 'server',
     analyzerHost: '127.0.0.1',
     analyzerPort: 8866
-};
+}
 const WebpackConfig = parentDirPath => outputPath => {
-    const output = outputPath ? outputPath : './deploy';
-    const smp = new SpeedMeasurePlugin();
+    const output = outputPath ? outputPath : './deploy'
+    const smp = new SpeedMeasurePlugin()
     return smp.wrap(
         merge(webpackBaseConfig(parentDirPath), {
             mode: 'development',
-            devtool: 'cheap-module-eval-source-map',
+            devtool: 'eval-cheap-module-source-map',
             entry: {},
             output: {
                 path: path.resolve(cwdPath, output),
@@ -34,8 +34,8 @@ const WebpackConfig = parentDirPath => outputPath => {
                 new webpack.NoEmitOnErrorsPlugin()
             ]
         })
-    );
-};
+    )
+}
 
 // export default WebpackConfig;
-module.exports = WebpackConfig;
+module.exports = WebpackConfig